--- a/lib/Core/Cache.php
+++ b/lib/Core/Cache.php
@@ -6,5 +6,9 @@ use Closure;
 
 interface Cache
 {
+    /**
+     * @param Closure(): mixed $closure
+     * @return mixed
+     */
     public function getOrSet(string $key, Closure $closure);
 }