--- a/lib/Core/Inference/FrameBuilder.php
+++ b/lib/Core/Inference/FrameBuilder.php
@@ -6,6 +6,7 @@ use Microsoft\PhpParser\FunctionLike;
 use Microsoft\PhpParser\Node;
 use Microsoft\PhpParser\Node\Expression\AnonymousFunctionCreationExpression;
 use Microsoft\PhpParser\Node\SourceFileNode;
+use Phpactor\WorseReflection\Core\Cache;
 use Phpactor\WorseReflection\Core\Inference\FrameBuilder\IncludeWalker;
 use Psr\Log\LoggerInterface;
 use RuntimeException;
@@ -30,7 +31,18 @@ final class FrameBuilder
      */
     private $walkers;
 
-    public static function create(DocBlockFactory $docblockFactory, SymbolContextResolver $symbolContextResolver, LoggerInterface $logger, array $walkers = [])
+    /**
+     * @var Cache
+     */
+    private $cache;
+
+    public static function create(
+        DocBlockFactory $docblockFactory,
+        SymbolContextResolver $symbolContextResolver,
+        LoggerInterface $logger,
+        Cache $cache,
+        array $walkers = []
+    ): self
     {
         $nameResolver = new FullyQualifiedNameResolver($logger);
         $walkers = array_merge([
@@ -44,16 +56,17 @@ final class FrameBuilder
             new IncludeWalker($logger),
         ], $walkers);
 
-        return new self($symbolContextResolver, $walkers);
+        return new self($symbolContextResolver, $walkers, $cache);
     }
 
     /**
      * @param FrameWalker[] $walkers
      */
-    public function __construct(SymbolContextResolver $symbolContextResolver, array $walkers)
+    public function __construct(SymbolContextResolver $symbolContextResolver, array $walkers, Cache $cache)
     {
         $this->symbolContextResolver = $symbolContextResolver;
         $this->walkers = $walkers;
+        $this->cache = $cache;
     }
 
     public function build(Node $node): Frame
@@ -61,39 +74,39 @@ final class FrameBuilder
         return $this->walkNode($this->resolveScopeNode($node), $node);
     }
 
-    private function walkNode(Node $node, Node $targetNode, Frame $frame = null)
+    private function walkNode(Node $node, Node $targetNode, ?Frame $frame = null): ?Frame
     {
-        if ($frame === null) {
-            $frame = new Frame($node->getNodeKindName());
-        }
+        $key = 'frame:'.spl_object_hash($targetNode);
 
-        if (null === $frame) {
-            throw new RuntimeException(
-                'Walk node was not intiated with a SouceFileNode, this should never happen.'
-            );
-        }
-
-        foreach ($this->walkers as $walker) {
-            if ($walker->canWalk($node)) {
-                $frame = $walker->walk($this, $frame, $node);
+        return $this->cache->getOrSet($key, function () use ($node, $targetNode, $frame) {
+            if ($frame === null) {
+                $frame = new Frame($node->getNodeKindName());
             }
-        }
 
-        foreach ($node->getChildNodes() as $childNode) {
-            if ($found = $this->walkNode($childNode, $targetNode, $frame)) {
-                return $found;
+            foreach ($this->walkers as $walker) {
+                if ($walker->canWalk($node)) {
+                    $frame = $walker->walk($this, $frame, $node);
+                }
             }
-        }
 
-        // if we found what we were looking for then return it
-        if ($node === $targetNode) {
-            return $frame;
-        }
+            foreach ($node->getChildNodes() as $childNode) {
+                if ($found = $this->walkNode($childNode, $targetNode, $frame)) {
+                    return $found;
+                }
+            }
 
-        // we start with the source node and we finish with the source node.
-        if ($node instanceof SourceFileNode) {
-            return $frame;
-        }
+            // if we found what we were looking for then return it
+            if ($node === $targetNode) {
+                return $frame;
+            }
+
+            // we start with the source node and we finish with the source node.
+            if ($node instanceof SourceFileNode) {
+                return $frame;
+            }
+
+            return null;
+        });
     }
 
 
@@ -141,4 +154,5 @@ final class FrameBuilder
 
         return $scopeNode;
     }
+
 }