--- a/lib/Core/ServiceLocator.php
+++ b/lib/Core/ServiceLocator.php
@@ -108,6 +108,7 @@ class ServiceLocator
             $this->docblockFactory,
             $this->symbolContextResolver,
             $this->logger,
+            $cache,
             $frameWalkers
         );
         $this->methodProviders = $methodProviders;