--- a/tests/Benchmarks/PhpUnitReflectClassBench.php
+++ b/tests/Benchmarks/PhpUnitReflectClassBench.php
@@ -9,7 +9,6 @@ use Phpactor\WorseReflection\Core\ClassName;
  * @Iterations(5)
  * @Revs(10)
  * @Warmup(1)
- * @OutputTimeUnit("milliseconds", precision=2)
  */
 class PhpUnitReflectClassBench extends BaseBenchCase
 {
@@ -24,7 +23,7 @@ class PhpUnitReflectClassBench extends BaseBenchCase
 
     /**
      * @Subject()
-     * @OutputTimeUnit("seconds", precision=2)
+     * @OutputTimeUnit("milliseconds", precision=2)
      * @Assert("variant.mode <= baseline.mode +/- 5%")
      */
     public function test_case_methods_and_properties(): void
@@ -45,7 +44,7 @@ class PhpUnitReflectClassBench extends BaseBenchCase
      *
      * @Subject()
      * @Revs(1)
-     * @OutputTimeUnit("seconds", precision=2)
+     * @OutputTimeUnit("milliseconds", precision=2)
      * @Assert("variant.mode <= baseline.mode +/- 5%")
      */
     public function test_case_method_frames(): void