Improves the frame processing time by 99.5% (for processing PHPunit frames

Small class:

```
0.09ms vs 22.71ms (±0.72%) -99.61% 
```

PHPUnit Test Case:

```
0.35ms vs 225.82ms (±2.42%) -99.85%
```

I'd hope this speeds up completion considerably especially in larger classes / frames.